+++ apps/commerce-api/src/main/java/com/loopers/application/payment/PaymentFacade.java
@@ -2,7 +2,6 @@ package com.loopers.application.payment;
 
 
 import com.loopers.domain.order.OrderModel;
-import com.loopers.domain.order.orderItem.OrderItemModel;
 import com.loopers.domain.payment.PaymentMethod;
 import com.loopers.domain.payment.PaymentModel;
 import com.loopers.domain.payment.PaymentStatus;
@@ -20,8 +19,6 @@ public class PaymentFacade {
   private final PaymentProcessor paymentProcessor;
   private final PaymentHistoryProcessor paymentHistoryProcessor;
   private final PaymentStrategyFactory paymentFactory;
-
-  private final StockProcessor stockProcessor;
 
   private final PaymentPublisher publisher;
   private final PaymentOrderProcessor processor;
@@ -62,11 +59,8 @@ public class PaymentFacade {
 
     // 성공인 경우
     // 재고 차감
-    for (OrderItemModel orderItem : orderModel.getOrderItems()) {
-      Long productId = orderItem.getProductId();
-      Long quantity = orderItem.getQuantity();
-      stockProcessor.decreaseStock(productId, quantity);
-    }
+    publisher.publish(orderModel.getOrderItems());
+
     paymentModel.done();
     publisher.publish(paymentModel.getOrderNumber());
     paymentHistoryProcessor.add(paymentModel, "결제가 완료되었습니다.");
