+++ apps/commerce-api/src/main/java/com/loopers/application/payment/PointUseHandler.java
@@ -1,24 +1,21 @@
 package com.loopers.application.payment;
 
-import com.loopers.domain.point.PointModel;
-import com.loopers.domain.point.PointRepository;
-import com.loopers.support.error.CoreException;
-import com.loopers.support.error.ErrorType;
-import java.math.BigInteger;
+import com.loopers.domain.point.PointUseCommand;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
+import org.springframework.transaction.event.TransactionPhase;
+import org.springframework.transaction.event.TransactionalEventListener;
 
+@Slf4j
 @Service
 @RequiredArgsConstructor
 public class PointUseHandler {
-  private final PointRepository pointRepository;
+  private final PaymentUsePointHandler handler;
 
-  @Transactional
-  public void use(String userId, BigInteger amount) {
-    PointModel hasPoint = pointRepository.getWithLock(userId).orElseThrow(
-        () -> new CoreException(ErrorType.BAD_REQUEST, "사용할 수 있는 포인트가 존재하지 않습니다.")
-    );
-    hasPoint.use(amount);
+  @TransactionalEventListener(phase = TransactionPhase.BEFORE_COMMIT)
+  public void use(PointUseCommand command) {
+    log.info("포인트 사용 이벤트 수신: {}", command);
+    handler.use(command.userId(), command.payment());
   }
 }
