+++ apps/commerce-api/src/main/java/com/loopers/domain/catalog/product/stock/StockDecreaseEvent.java
@@ -1,0 +1,7 @@
+package com.loopers.domain.catalog.product.stock;
+
+import com.loopers.domain.order.orderItem.OrderItemModel;
+import java.util.List;
+
+public record StockDecreaseEvent(List<OrderItemModel> orderItems) {
+}
