+++ apps/commerce-api/src/main/java/com/loopers/domain/point/PointUseCommand.java
@@ -1,0 +1,6 @@
+package com.loopers.domain.point;
+
+import java.math.BigInteger;
+
+public record PointUseCommand(String userId, BigInteger payment) {
+}
