+++ apps/commerce-api/src/main/java/com/loopers/infrastructure/payment/PaymentCorePublisher.java
@@ -3,8 +3,13 @@ package com.loopers.infrastructure.payment;
 import com.loopers.application.payment.PaymentPublisher;
 import com.loopers.data_platform.PlatformSendEvent;
 import com.loopers.data_platform.UserTrackingData;
+import com.loopers.domain.catalog.product.stock.StockDecreaseEvent;
+import com.loopers.domain.order.orderItem.OrderItemModel;
 import com.loopers.domain.payment.PaymentOrderConfirmCommand;
+import com.loopers.domain.point.PointUseCommand;
+import java.math.BigInteger;
 import java.time.ZonedDateTime;
+import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.ApplicationEventPublisher;
@@ -25,6 +30,18 @@ public class PaymentCorePublisher implements PaymentPublisher {
     publisher.publishEvent(new PlatformSendEvent(
         "결제", paymentId, result
     ));
+  }
+
+  // 포인트
+  @Override
+  public void publish(String userId, BigInteger payment) {
+    publisher.publishEvent(new PointUseCommand(userId, payment));
+  }
+
+  // 재고
+  @Override
+  public void publish(List<OrderItemModel> orderItems) {
+    publisher.publishEvent(new StockDecreaseEvent(orderItems));
   }
 
   @Override
@@ -52,5 +69,7 @@ public class PaymentCorePublisher implements PaymentPublisher {
         )
     );
   }
+
+
 
 }
