+++ apps/commerce-api/src/test/java/com/loopers/application/payment/OrderStockProcessorTest.java
@@ -2,6 +2,7 @@ package com.loopers.application.payment;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import com.loopers.application.catalog.product.StockProcessor;
 import com.loopers.domain.catalog.product.stock.StockModel;
 import com.loopers.domain.catalog.product.stock.StockRepository;
 import com.loopers.infrastructure.catalog.product.ProductJpaRepository;
